requirements.txt
flask
google-generativeai
pillow
pytesseract
gunicorn
